js側の修正
