+++ app/javascript/add.js
@@ -1,0 +1,12 @@
+function add(){
+  const buttonPlus = document.getElementById("button_plus");
+  
+  buttonPlus.addEventListener("click", () => {
+    const form = document.querySelectorAll(".ingredients");
+    const formNumber = form.length - 1;
+    form[formNumber].insertAdjacentHTML("afterend", ingredient);
+  });
+
+};
+
+window.addEventListener("load", add)
